refactor: minor styling changes to folder structure, exports, types
