+++ packages/idempotency/src/config/ConfigServiceInterface.ts
@@ -1,0 +1,13 @@
+interface ConfigServiceInterface {
+
+  get(name: string): string
+
+  getServiceName(): string
+
+  getFunctionName(): string
+
+}
+
+export {
+  ConfigServiceInterface
+};
