+++ packages/idempotency/src/persistence/PersistenceLayerInterface.ts
@@ -1,7 +1,6 @@
 import { IdempotencyRecord } from './IdempotencyRecord';
 
 interface PersistenceLayerInterface {
-  configure(functionName: string): void
   saveInProgress(data: unknown): Promise<void>
   saveSuccess(data: unknown, result: unknown): Promise<void>
   deleteRecord(data: unknown): Promise<void>
