+++ packages/idempotency/src/types/IdempotencyOptions.ts
@@ -1,4 +1,4 @@
-import { PersistenceLayer } from './persistence/PersistenceLayer';
+import { PersistenceLayer } from '../persistence/PersistenceLayer';
 
 type IdempotencyOptions = {
   dataKeywordArgument: string
