+++ packages/idempotency/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './AnyFunction';
 export * from './IdempotencyRecordStatus';
+export * from './PersistenceLayer';
